services/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.27.0
websockets==13.1
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
apscheduler==3.10.4
pytz==2024.1
numpy==1.26.4
pandas==2.2.2
ib_insync==0.9.86
torch==2.4.0+cpu
scikit-learn==1.5.1
yfinance==0.2.41
joblib==1.4.2
ta==0.11.0
vaderSentiment==3.3.2
newsapi-python==0.2.7

--extra-index-url https://download.pytorch.org/whl/cpu
